requirements.txt
# QA Portal Dependencies - Supabase Filter Fix Deployment
# CACHE_BUSTER: CSV_DESCRIPTION_FIX_sept5_2025_f82b2c4
streamlit>=1.28.1
pandas>=2.0.0
pyarrow>=14.0.0
openai>=1.0.0
supabase>=2.0.0
pyairtable>=2.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
outscraper>=3.0.0
aiohttp>=3.8.0
toml>=0.10.0
pandera>=0.17.0
plotly>=5.20.0
fpdf2>=2.7.0
